refactor(backend): move towards workspace layout
