+++ src-tauri/crates/sjmcl-macros/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "sjmcl-macros"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+
+[lib]
+proc-macro = true
+
+[dependencies]
+proc-macro2.workspace = true
+quote.workspace = true
+syn.workspace = true
