--- src-tauri/libs/partial_derive/.gitignore
@@ -1,1 +1,0 @@
-target/
