--- src-tauri/libs/serialize_skip_none_derive/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "serialize_skip_none_derive"
-version = "0.1.0"
-edition = "2024"
-
-[lib]
-proc-macro = true
-
-[dependencies]
-syn = {version = "2", features = ["full"]}
-quote = "1"
-proc-macro2 = "1"
