+++ src-tauri/src/launcher_config/models.rs
@@ -5,8 +5,8 @@ use crate::storage::Storage;
 use crate::utils::string::snake_to_camel_case;
 use crate::utils::sys_info;
 use crate::{APP_DATA_DIR, EXE_DIR, IS_PORTABLE};
-use partial_derive::Partial;
 use serde::{Deserialize, Serialize};
+use sjmcl_macros::Partial;
 use smart_default::SmartDefault;
 use std::path::PathBuf;
 use strum_macros::Display;
